+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **_Level Stats Format_** settings
   - Removed _Kills % in Stats display_ setting in favor of them
+- **Minimap zooming** (keyboard only)
 
 ## Changes